circular dependencies and modularizing const/let cleaning
