+++ server/src/utils/context.js
@@ -14,7 +14,7 @@ should FE have an active context indicator?
 
 Take in a notebook.
 */
-const main = require("./engine");
+// const main = require("./engine");
 
 const contextStore = {
   "111": {
@@ -22,6 +22,43 @@ const contextStore = {
     lastUpdate: 1687812247076,
     context: {}
   }
+}
+
+const variableMap = {};
+const cellsRun = {};
+const flagged = {};
+const changedVariables = {}
+
+const updateVariableMap = (variable, value) => {
+  variableMap[variable] = value;
+}
+
+const updateCellsRun = (cellId, value) => {
+  cellsRun[cellId] = value
+}
+
+const updateFlagged = (variable) => {
+  flagged[variable] = true;
+}
+
+const updateChangedVariables = (variable, cellId) => {
+  changedVariables[variable] = `${variable}${cellsRun[cellId]}` 
+}
+
+const getVariableMap = () => {
+  return variableMap;
+}
+
+const getCellsRun = () => {
+  return cellsRun;
+}
+
+const getFlagged = () => {
+  return flagged;
+}
+
+const getChangedVariables = () => {
+  return changedVariables;
 }
 
 const loadContext = (notebookId) => {
@@ -44,13 +81,13 @@ const resetContext = (notebookId) => {
   contextStore[notebookId].active = false;
   contextStore[notebookId].context = {};
   //resets the in-memory state of context in engine.js
-  for (var member in main.variableMap) delete main.variableMap[member];
-  for (var member in main.cellsRun) delete main.cellsRun[member];
-  for (var member in main.flagged) delete main.flagged[member];
+  for (var member in variableMap) delete variableMap[member];
+  for (var member in cellsRun) delete cellsRun[member];
+  for (var member in flagged) delete flagged[member];
 }
 
 const updateContextWrapper = (notebookId) => {
   contextStore[notebookId].lastUpdate = Date.now();
 }
 
-module.exports = { loadContext, updateContextWrapper, resetContext};
+module.exports = { loadContext, updateContextWrapper, resetContext, getVariableMap, getCellsRun, getFlagged, getChangedVariables, updateVariableMap, updateCellsRun, updateFlagged, updateChangedVariables};
